% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_plot.R
\name{clean_rose_inputs}
\alias{clean_rose_inputs}
\title{Takes in rose chart inputs from the UI and outputs a
tibble to be used in the ggrose function}
\usage{
clean_rose_inputs(
  personnel,
  dispersal,
  duration,
  screening,
  socialdist,
  masks,
  localpop,
  localcov,
  icu
)
}
\description{
Takes in rose chart inputs from the UI and outputs a
tibble to be used in the ggrose function
}
