% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_plot.R
\name{convert_rating}
\alias{convert_rating}
\title{Converts ratings to numeric values, or numeric values to ratings}
\usage{
convert_rating(input)
}
\description{
Converts ratings to numeric values, or numeric values to ratings
}
