% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{die_out_fun}
\alias{die_out_fun}
\title{Die Out Function}
\usage{
die_out_fun(x, m = 0.02)
}
\arguments{
\item{x}{Number of infected individuals}

\item{m}{Die out threshold}
}
\description{
Die Out Function
}
