% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_plot.R
\name{ggrose}
\alias{ggrose}
\title{GGRose}
\usage{
ggrose(
  personnel,
  dispersal,
  duration,
  screening,
  socialdist,
  masks,
  localpop,
  localcov,
  icu,
  additional_risk,
  chart_num
)
}
\value{
A ggplot object
}
\description{
A graphical wrapper function that takes inputs from the Rose Chart
UI page and resturns a graphical output of thos inputs.
}
