% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highline}
\alias{highline}
\title{2017 Highline Arizona fire information}
\format{
A tibble data frame with 24 rows and 9 variables
}
\usage{
highline
}
\description{
A dataset containing summaries of resource mobilization to and from
the 2017 Highline fire.
}
\keyword{datasets}
