% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{Launches the CIRA Dashboard as a shiny app}
\usage{
launch_application(...)
}
\arguments{
\item{...}{Additional args for \code{shiny::runApp}}
}
\description{
Launches the CIRA Dashboard as a shiny app
}
