% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_ciraDash.R
\name{launch_ciraDash}
\alias{launch_ciraDash}
\title{Launches the CIRA Dashboard as a shiny app}
\usage{
launch_ciraDash(...)
}
\arguments{
\item{...}{Additional args for \code{shiny::runApp}}
}
\description{
Launches the CIRA Dashboard as a shiny app
}
