% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lolo_peak}
\alias{lolo_peak}
\title{2017 Lolo Peak Montana fire information}
\format{
A tibble data frame with 80 rows and 9 variables
}
\usage{
lolo_peak
}
\description{
A dataset containing summaries of resource mobilization to and from
the 2017 Lolo Peak Fire
}
\keyword{datasets}
