% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{make_entry_exit}
\alias{make_entry_exit}
\title{Make Entry Exit Data}
\usage{
make_entry_exit(t)
}
\arguments{
\item{t}{Numeric input, a length of time}
}
\description{
This is data generated as a stand in for real ROSS data.
If there is no Entry/Exit data given to the SEIR Setup function
it will use this function to create data.
}
