% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_plot.R
\name{rationale_box}
\alias{rationale_box}
\title{Creates a text box for rationales}
\usage{
rationale_box(name)
}
\arguments{
\item{name}{Name of a rose chart input}
}
\description{
This function gets used in ui.R to create text input boxes for rationales
}
\examples{
rationale_box("duration")
}
