% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ross.R
\name{read_ross}
\alias{read_ross}
\title{Read ROSS}
\usage{
read_ross(path)
}
\arguments{
\item{path}{File path}
}
\description{
Takes in summarized ROSS data, cleans it up and returns a
data frame that can be used by the SEIR diff-eq model
}
