% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{SEIR_entry_exit}
\alias{SEIR_entry_exit}
\title{SEIR Epi Model wrapper}
\usage{
SEIR_entry_exit(inputs, equation, .die_out = FALSE)
}
\arguments{
\item{inputs}{A data frame that was created with the SEIR_setup function}

\item{equation}{The desired set of differential equations}

\item{.die_out}{Logical, implement stochastic die out}
}
\description{
This function wraps the solver and formats output data for plotting and processing
}
