% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{SEIR_eqn}
\alias{SEIR_eqn}
\title{Defining the SEIR equations}
\usage{
SEIR_eqn(time, state, parameters)
}
\arguments{
\item{time}{Length of time}

\item{state}{Initial states}

\item{parameters}{Parameters that change}
}
\description{
A series of differential equations used by the SEIR_entry_exit function
to run the model
}
