% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{SEIR_setup}
\alias{SEIR_setup}
\title{SEIR Setup}
\usage{
SEIR_setup(
  t = 70,
  R0 = 2.68,
  De = 5,
  Di = 3,
  N.init = 10,
  I.init = 1,
  eir = 0.1,
  varying
)
}
\arguments{
\item{t}{Simulation duration, in days}

\item{R0}{Basic reproduction number}

\item{De}{Incubation period, in days}

\item{Di}{Infectious period, in days}

\item{N.init}{Initial population}

\item{I.init}{Initial number of infected individuals}

\item{eir}{Entry Infection Rate}

\item{varying}{Parameters that vary over time}
}
\description{
Constructs a parameter data frame for use in the SEIR model function
}
