% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{show_selection}
\alias{show_selection}
\title{Show selection}
\usage{
show_selection(
  personnel,
  dispersal,
  duration,
  screening,
  socialdist,
  masks,
  localpop,
  localcov,
  icu,
  additional_risk
)
}
\description{
Change ratings back into selections
}
