% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tank_hollow}
\alias{tank_hollow}
\title{2017 Tank Hollow fire information}
\format{
A tibble data frame with 80 rows and 9 variables
}
\usage{
tank_hollow
}
\description{
A dataset containing summaries of resource mobilization to and from
the 2017 Tank Hollow fire.
}
\keyword{datasets}
