% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_plot.R
\name{tool_tip}
\alias{tool_tip}
\title{Creates a tooltip}
\usage{
tool_tip(variable, text)
}
\arguments{
\item{variable}{The name/title of the tooltip}

\item{text}{The hover over text that appears over the title}
}
\description{
This function gets used in the server.R to create tooltips for the different rose chart inputs
}
