% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose_plot.R
\name{convert_rating}
\alias{convert_rating}
\title{Converts ratings to numeric or vice versa}
\usage{
convert_rating(input)
}
\arguments{
\item{input}{A character vector or numeric input}
}
\value{
The character input \code{'low'} as the numeric output \code{1}
The numeric input \code{1} as the character output \code{'Low'}
}
\description{
Converts ratings to numeric or vice versa
}
